Add 'Reset' button. General refactoring

+++ __init__.py
@@ -1,7 +1,6 @@
 import tkinter as tk
 import os
 from tkinter import filedialog, messagebox
-from typing import Tuple, Any
 
 
 class MainApplication(tk.Tk):
@@ -26,9 +25,9 @@ class MainApplication(tk.Tk):
                 self.source_files_box.itemconfig(index, {'fg': 'black'})
                 target_box.itemconfig(index, {'fg': 'black'})
             index += 1
-        ceva = ('moveto', self.scroll_bar_position[0])
-        self.source_files_box.yview(*ceva)
-        self.result_files_box.yview(*ceva)
+        scroll_position = ('moveto', self.scroll_bar_position[0])
+        self.source_files_box.yview(*scroll_position)
+        self.result_files_box.yview(*scroll_position)
 
     def rename_all_files(self, new_name):
         if self.rename_radio_value.get() == 2:
@@ -51,16 +50,34 @@ class MainApplication(tk.Tk):
         self.output_message_label.configure(text="")
 
     def save_button_handler(self):
-        current_dir = self.path_input_text_variable.get()
-        renamed_files = 0
+        files_to_rename = 0
         for item in self.files_map:
             if item != self.files_map[item]:
-                src = current_dir + "\\" + item
-                dst = current_dir + "\\" + self.files_map[item]
-                os.rename(src, dst)
-                renamed_files += 1
-        self.output_message_label.configure(text=str(renamed_files) + " files renamed!")
-        self.build_files_box()
+                files_to_rename += 1
+        if files_to_rename > 0:
+            if messagebox.askyesno('Rename Confirmation',
+                                   'Do you really want to rename ' + str(files_to_rename) + ' files?'):
+                files_not_renamed = 0
+                current_dir = self.path_input_text_variable.get()
+                renamed_files = 0
+                for item in self.files_map:
+                    if item != self.files_map[item]:
+                        src = current_dir + "\\" + item
+                        dst = current_dir + "\\" + self.files_map[item]
+                        try:
+                            os.rename(src, dst)
+                            renamed_files += 1
+                        except FileExistsError:
+                            files_not_renamed += 1
+                self.output_message_label.configure(text=str(renamed_files) + " files successfully renamed!")
+                self.build_files_box()
+                if files_not_renamed > 0:
+                    self.error_message_label.configure(text="Cannot rename " + str(files_not_renamed)
+                                                            + " file(s). File already exists!")
+            else:
+                self.reset_messages()
+        else:
+            self.error_message_label.configure(text="Noting to rename!")
 
     def open_path_button(self):
         # Enter Path section
@@ -141,9 +158,8 @@ class MainApplication(tk.Tk):
 
         self.show_items(self.files_map.values(), self.result_files_box, True)
 
-
     def build_files_box(self):
-        # self.reset_messages()
+        self.reset_messages()
         # Build a dictionary(originalFilename -> updatedFilename) from provided input
         try:
             self.files_map.clear()
@@ -166,6 +182,9 @@ class MainApplication(tk.Tk):
             self.source_files_box.delete(0, tk.END)
             self.result_files_box.delete(0, tk.END)
 
+        except FileExistsError:
+            self.error_message_label.configure(text="Cannot create a file when that file already exists")
+
     def build_input_frame(self):
         def enter_handler(event):
             self.build_files_box()
@@ -180,46 +199,52 @@ class MainApplication(tk.Tk):
             self.rename_all_files(new_name)
 
         # Enter Path section
-        path_label = tk.Label(self.inputFrame, text="Enter Path")
+        path_label = tk.Label(self.inputFrame, text="Enter Path: ")
         path_label.grid(row=0, column=0, pady=(10, 0), padx=(10, 0))
         self.path_input_text_variable = tk.StringVar()
         self.path_input = tk.Entry(self.inputFrame, width=30, textvariable=self.path_input_text_variable)
         self.path_input.bind('<Return>', enter_handler)
         self.path_input.grid(row=0, column=1, pady=(10, 0), padx=(0, 0))
-        tk.Button(self.inputFrame, text="browse", command=self.open_path_button, height=1).grid(row=0, column=2, pady=(10, 0), padx=(0, 0))
+        tk.Button(self.inputFrame, text="browse", command=self.open_path_button, height=1).grid(row=0, column=2,
+                                                                                                pady=(10, 0),
+                                                                                                padx=(0, 0))
 
         # Enter Name section
-        tk.Label(self.inputFrame, text="Enter Name").grid(row=0, column=3, pady=(10, 0), padx=(365, 0))
+        tk.Label(self.inputFrame, text="Enter Name:").grid(row=0, column=3, pady=(10, 0), padx=(365, 0), sticky='e')
         self.name_input_text_variable = tk.StringVar()
         name_input = tk.Entry(self.inputFrame, width=30, textvariable=self.name_input_text_variable)
-        name_input.grid(row=0, column=4, pady=(10, 0), padx=(10, 0))
+        name_input.grid(row=0, column=4, columnspan=2, pady=(10, 0), padx=(10, 0))
         name_input.bind('<KeyRelease>', key_pressed_handler)
 
         # Rename all Radio Buttons
         self.rename_radio_value = tk.IntVar()
         tk.Label(self.inputFrame, text="Rename:").grid(row=1, column=3, pady=(0, 0), padx=(365, 0), sticky='e')
         tk.Radiobutton(self.inputFrame, text="All", variable=self.rename_radio_value,
-                       value=1, command=lambda: select_handler()).grid(row=1, column=4, pady=(0, 0), padx=(10, 0), sticky='w')
+                       value=1, command=lambda: select_handler()) \
+            .grid(row=1, column=4, pady=(0, 0), padx=(10, 0), sticky='w')
         tk.Radiobutton(self.inputFrame, text="Selected", variable=self.rename_radio_value,
-                       value=2, command=lambda: select_handler()).grid(row=2, column=4, pady=(0, 0), padx=(10, 0), sticky='w')
+                       value=2, command=lambda: select_handler()) \
+            .grid(row=1, column=5, pady=(0, 0), padx=(10, 0), sticky='w')
         self.rename_radio_value.set(1)
 
-        tk.Button(self.inputFrame, text="Step-", command=lambda: self.step_button('step', False), height=1)\
-            .grid(row=3, column=5, padx=(0, 0), pady=(0, 0))
-        tk.Button(self.inputFrame, text="Step+", command=lambda: self.step_button('step', True), height=1)\
-            .grid(row=3, column=6, padx=(0, 0), pady=(0, 0))
-        tk.Button(self.inputFrame, text="cl-", command=lambda: self.step_button('_cl', False), height=1)\
-            .grid(row=3, column=7, padx=(10, 0), pady=(0, 0))
-        tk.Button(self.inputFrame, text="cl+", command=lambda: self.step_button('_cl', True), height=1)\
-            .grid(row=3, column=8, padx=(0, 0), pady=(0, 0))
-        tk.Button(self.inputFrame, text="en-", command=lambda: self.step_button('_en', False), height=1)\
-            .grid(row=3, column=9, padx=(10, 0), pady=(0, 0))
-        tk.Button(self.inputFrame, text="en+", command=lambda: self.step_button('_en', True), height=1)\
-            .grid(row=3, column=10, padx=(0, 0), pady=(0, 0))
-        tk.Button(self.inputFrame, text="i-", command=lambda: self.step_button('_i', False), height=1)\
-            .grid(row=3, column=11, padx=(10, 0), pady=(0, 0))
-        tk.Button(self.inputFrame, text="i+", command=lambda: self.step_button('_i', True), height=1)\
-            .grid(row=3, column=12, padx=(0, 0), pady=(0, 0))
+        tk.Button(self.inputFrame, text="Remove 'Copy'", command=lambda: self.remove_copy_button(), height=1) \
+            .grid(row=2, column=4, padx=(0, 0), pady=(0, 0))
+        tk.Button(self.inputFrame, text="Step-", command=lambda: self.step_button('step', False), height=1) \
+            .grid(row=2, column=6, padx=(0, 0), pady=(0, 0))
+        tk.Button(self.inputFrame, text="Step+", command=lambda: self.step_button('step', True), height=1) \
+            .grid(row=2, column=7, padx=(0, 0), pady=(0, 0))
+        tk.Button(self.inputFrame, text="cl-", command=lambda: self.step_button('_cl', False), height=1) \
+            .grid(row=2, column=8, padx=(10, 0), pady=(0, 0))
+        tk.Button(self.inputFrame, text="cl+", command=lambda: self.step_button('_cl', True), height=1) \
+            .grid(row=2, column=9, padx=(0, 0), pady=(0, 0))
+        tk.Button(self.inputFrame, text="en-", command=lambda: self.step_button('_en', False), height=1) \
+            .grid(row=2, column=10, padx=(10, 0), pady=(0, 0))
+        tk.Button(self.inputFrame, text="en+", command=lambda: self.step_button('_en', True), height=1) \
+            .grid(row=2, column=11, padx=(0, 0), pady=(0, 0))
+        tk.Button(self.inputFrame, text="i-", command=lambda: self.step_button('_i', False), height=1) \
+            .grid(row=2, column=12, padx=(10, 0), pady=(0, 0))
+        tk.Button(self.inputFrame, text="i+", command=lambda: self.step_button('_i', True), height=1) \
+            .grid(row=2, column=13, padx=(0, 0), pady=(0, 0))
 
     def build_output_frame(self):
         def on_single_click_release(event):
@@ -242,47 +267,51 @@ class MainApplication(tk.Tk):
             self.result_files_scrollbar.set(*args)
             self.scroll_bar_position = (args[0], args[1])
 
-
         def yview(*args):
             self.source_files_box.yview(*args)
             self.result_files_box.yview(*args)
 
         # Before box - displaying original filenames
+        tk.Label(self.outputFrame, text="Before:").grid(row=0, column=0, pady=(10, 0), padx=(10, 0), sticky='w')
         self.source_files_box = tk.Listbox(self.outputFrame, width=100, height=45, selectmode=tk.EXTENDED,
                                            activestyle='none')
-        self.source_files_box.grid(row=0, column=0, pady=10, padx=(10, 0))
+        self.source_files_box.grid(row=1, column=0, pady=10, padx=(10, 0))
         self.source_files_box.bind('<Double-Button-1>', self.on_double_click)
         self.source_files_box.bind('<ButtonRelease-1>', on_single_click_release)
         self.source_files_scrollbar = tk.Scrollbar(self.outputFrame, orient="vertical")
-        self.source_files_scrollbar.grid(row=0, column=1, pady=10, sticky="nsw")
+        self.source_files_scrollbar.grid(row=1, column=1, pady=10, sticky="nsw")
         self.source_files_scrollbar.config(command=yview)
         self.source_files_box.config(yscrollcommand=yscroll1)
         self.source_files_box.configure(exportselection=False)
 
         # '->' label between the 2 boxes
         # TODO: add a custom image instead of plain text
-        tk.Label(self.outputFrame, text="->", anchor="center").grid(row=0, column=2, padx=(10, 0), sticky='nsew')
+        tk.Label(self.outputFrame, text="->", anchor="center").grid(row=1, column=2, padx=(10, 0), sticky='nsew')
 
         # After box - displaying updated filenames
+        tk.Label(self.outputFrame, text="After:").grid(row=0, column=3, pady=(10, 0), padx=(10, 0), sticky='w')
         self.result_files_box = tk.Listbox(self.outputFrame, width=100, height=45,
                                            selectmode=tk.SINGLE, selectbackground='white',
                                            selectforeground='black',
                                            activestyle='none')
-        self.result_files_box.grid(row=0, column=3, pady=10, padx=(10, 0))
+        self.result_files_box.grid(row=1, column=3, pady=10, padx=(10, 0))
         self.result_files_box.bind('<Double-Button-1>', self.on_double_click)
         self.result_files_scrollbar = tk.Scrollbar(self.outputFrame, orient="vertical")
-        self.result_files_scrollbar.grid(row=0, column=4, pady=10, sticky="nsw")
+        self.result_files_scrollbar.grid(row=1, column=4, pady=10, sticky="nsw")
         self.result_files_scrollbar.config(command=yview)
         self.result_files_box.config(yscrollcommand=yscroll2)
 
     def build_messages_frame(self):
-        self.output_message_label = tk.Label(self.messagesFrame, width=165, text="", anchor="w")
+        self.output_message_label = tk.Label(self.messagesFrame, width=145, text="", anchor="w")
         self.output_message_label.grid(row=0, column=0, padx=10, sticky='w')
 
-        self.error_message_label = tk.Label(self.messagesFrame, width=165, text="", anchor="w")
+        self.error_message_label = tk.Label(self.messagesFrame, width=145, text="", anchor="w")
         self.error_message_label.grid(row=1, column=0, padx=10, sticky='w')
-        tk.Button(self.messagesFrame, text="Save", command=self.save_button_handler)\
-            .grid(row=0, column=2, pady=(10, 0), padx=(10, 10), sticky='e')
+        tk.Button(self.messagesFrame, text="Refresh", command=self.build_files_box, width=15) \
+            .grid(row=0, column=2, pady=(10, 0), padx=(0, 0))
+
+        tk.Button(self.messagesFrame, text="Save", command=self.save_button_handler, width=15) \
+            .grid(row=0, column=3, pady=(10, 0), padx=(5, 0))
 
     def on_double_click(self, evt):
         # Note here that Tkinter passes an event object to onselect()
